new code added

+++ ToDoTask/Views/ListView.swift
@@ -10,23 +10,29 @@ import SwiftUI
 struct ListView: View {
     @EnvironmentObject var listViewModel : ListViewModel
     var body: some View {
-            List{
-                ForEach(listViewModel.items){ item in
-                    rowList(value:item)
-                        .onTapGesture {
-                            withAnimation(.linear){
-                                listViewModel.updateItem(item: item)
+        ZStack{
+            if listViewModel.items.isEmpty{
+                Text("is empty")
+            }else{
+                List{
+                    ForEach(listViewModel.items){ item in
+                        rowList(value:item)
+                            .onTapGesture {
+                                withAnimation(.linear){
+                                    listViewModel.updateItem(item: item)
+                                }
                             }
-                        }
+                    }
+                    .onDelete(perform:listViewModel.deleteItem)
+                    .onMove(perform:listViewModel.moveItem)
                 }
-                .onDelete(perform:listViewModel.deleteItem)
-                .onMove(perform:listViewModel.moveItem)
-            }/*Navigation Bar items*/
+            }
+        }
+           /*Navigation Bar items*/
             //.listStyle(PlainListStyle())
             .navigationBarTitle("To Do App")
             .navigationBarItems(leading: EditButton(), trailing: NavigationLink("Add", destination:addNewTask()))
-            }
- 
+  }
 }
 struct ListView_Previews: PreviewProvider {
     static var previews: some View {
